Added communication between Lobby and Proxy
The Lobby will now send the new best time to the proxy, and it will return the HashMap containing the best parkour times. All communication uses PluginMessages from BungeeCord
The /bestparkourtimes command was also added. It will always return an updated list of best parkour times, queried to the Proxy before display
The best times are not yet saved in a database on the Proxy side, but it will soon